pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "palimpsest"
version = "0.1.0"
description = "Personal journal metadata management and PDF compilation system"
authors = [{ name = "Sofía F." }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
  "sqlalchemy>=2.0.43",
  "alembic>=1.16.5",
  "click>=8.3.0",
  "pyyaml>=6.0.2",
  "pypandoc>=1.15",
  "ftfy>=6.3.1",
  "textstat>=0.7.10",
  "markupsafe>=3.0.2",
  "mako>=1.3.10",
  "greenlet>=3.2.2",
  "jinja2>=3.1.0",
  "markdown-it-py>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.1.0", "ruff>=0.2.0"]

[project.scripts]
plm = "dev.pipeline.cli:cli"
jsearch = "dev.search.cli:cli"
metadb = "dev.database.cli:cli"
validate = "dev.validators.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["dev*"]

[tool.ruff]
line-length = 88
target-version = "py310"
